App.Draw.Render

Documentation

playerDraw

Arguments

:: Int
the number of the paddle. (1 for the player paddle, anything else for the AI paddle)

-> Float
the x-coordinate of the center of the paddle

-> Float
a number used to generate the inner color of the paddle

-> Picture
the resulting picture

Draws a paddle for a specified player according to its position.

ballDraw

Arguments

:: Point
the (x,y)-coordinates of the ball

-> Float
a number used to generate the inner color of the ball

-> Picture
the resulting picture

Draws the ball according to its position.

displayTime

Arguments

:: Float
a number used to generate the color of the text

-> World
the current world

-> Picture
the resulting picture

Draws the game time of a world.

victory

Arguments

:: Float
a number used to generate the color of the text

-> World
the current world

-> Picture
the resulting picture

Draws a victory screen for a given world. Displays the total time of the game.

defeat

Arguments

:: Float
a number used to generate the color of the text

-> World
the current world

-> Picture
the resulting picture

Draws a defeat screen for a given world. Displayes the total time of the game.

leaderboardSort :: [Record] -> [Record]

Sorts a list of game records.

leaderboard :: IO Picture

The top-10 scores of all time as a picture wrapped in IO context.

idle

Arguments

:: Color
the color of the title text

-> IO Picture
the resulting picture to be used in an IO context.

Generates an idle-screen wrapped in IO context with a given color.

pause

Arguments

:: Color
the color of the title text

-> IO Picture
the resulting picture to be used in an IO context.

Generates a pause-screen wrapped in IO context with a given color.

worldDraw

Arguments

:: World
the current world

-> IO Picture
the resulting picture to be used in an IO context.

A picture of a world wrapped in IO context.