App.Logic

Documentation

data Game

A game value. It holds all necessary data to calculate the next game state depending on time.

Constructors

Game

Fields

p1 :: Float

x-coordinate of the player-controlled paddle

p2 :: Float

x-coordinate of the computer-controlled paddle

ball :: Point

the coordinates of the ball

dir :: Float

the direction in which the ball is moving in radians

vel :: Float

the velocity of the ball

data Result

Describes the result and/or state of the current world.

Constructors

Ongoing
for games in progress

Player
for finished games won by the player

AI
for finished games won by the AI

Idle
for games that have not been started yet

Paused
for paused games

data World

Describes the world of the game. Holds the game data and the score, time and result of the game.

Constructors

World

Fields

game :: Game

the current game state

p1Score :: Int

player's score

p2Score :: Int

AI's score

result :: Result

the result/state of the current game

time :: Float

the playtime (only includes active playing)

idleTime :: Float

the time of inactivity (only when the game is not being played)

type Record = String

Used for storing the time and result of a finished game.

startWorld :: World

The default world of the game. By default the game has not been started yet, the paddles and the ball are centered, and the ball direction is 0.72 (in radians)

recordUncurry

Arguments

:: Record
the record string (minutes:seconds:miliseconds player:ai)

-> (Float, Int)
the resulting time-player pair

Returns a pair of time in seconds and the player's score of a given record

paddleAngle

Arguments

:: Float
the x-coordinate of the paddle

-> Float
the x-coordinate of the ball

-> Float
the resulting angle in radians

Returns the bounce angle of a ball hitting a paddle.

normalize

Arguments

:: Float
the coordinate to be normalized

-> Float
the resulting coordinate

Returns a normalized coordinate for collision-avoidance purposes.

moveBall

Arguments

:: Float
time passed in seconds

-> World
the world to update

-> World
the resulting world

Moves the ball in a set direction for a given time accounting for collisions.

scoreCheck

Arguments

:: World
the current world

-> World
the world with updated score/field in case someone scored

Changes the score and the game state if any player scores. Also changes the speed of the ball for each point scored by the player.

aiMove

Arguments

:: World
the current world

-> World
the world with updated paddle position

Moves the second player's paddle according to where the ball is. The bigger the player's score, the faster the paddle moves.

gameOver

Arguments

:: World
the current world

-> World
the world with updated result.

Checks if the game is over and changes the result accordingly. The game ends once someone has scored 7 points.

updateWorld

Arguments

:: Float
time passed in seconds

-> World
the current world

-> IO World
the updated world wrapped in IO context

Updates the world according to the time passed.

eventHandler

Arguments

:: Event
the event to be processed

-> World
the current world

-> IO World
the updated world wrapped in IO context

Updates the world according to an input event.

timeAsText

Arguments

:: World
the current world

-> String
the time of active playing

Returns the time of the game in a (minutes:seconds:miliseconds) format.

scoreAsText

Arguments

:: World
the current world

-> String
the current score

Returns the current score of the world in a (player score):(AI score) format.