Lib

Description

Main game module

Documentation

updateSeed :: PongGame -> PongGame

Function that sets game seed

window :: Display

Game window declaration, uses values of width and height from Types module

background :: Color

Sets the background color

render :: PongGame -> Picture

Renders current view of the game

initialState :: PongGame

Sets the game to starting values

moveBall :: Float -> PongGame -> PongGame

Function that draws the ball during the game

type Radius = Float

Type used to store ball radius

type Position = (Float, Float)

Type used to store position

wallCollision :: Position -> Radius -> Bool

Function that checks if ball collides with wall

paddleCollision :: Position -> Float -> Float -> Radius -> Bool

Function that checks if ball collides with paddle

ballBounce :: PongGame -> PongGame

Function that switches the ball direction after collision

playerMovement :: PongGame -> Int -> PongGame

Function that gets player move

computerMovement :: PongGame -> PongGame

Function that calculate computer move

movePlayer :: PongGame -> PongGame

Function that moves player

resetBall :: PongGame -> PongGame

Function that resets ball position after collision with right or left wall

update :: Float -> PongGame -> PongGame

Function that updates view

handleKeys :: Event -> PongGame -> PongGame

Function that takes player's input

doPong :: IO ()

Function that runs the game